oke gotta refactor workflow now
