+++ src/backend_new/main.py
@@ -114,134 +114,6 @@ class Analyzer:
         ...
 
     # endregion
-    def process_song(self) -> None:
-        """
-        Processes a song, allows the user to choose which song to process
-        """
-        # region helper functions
-        def update_song_data() -> None:
-            """Updates the chosen song's data in the song_data variable."""
-            nonlocal song_context
-            song_context.json_song_data = read_json_file(song_context.json_file_path)
-        # endregion
-
-        # region asks user which song to process
-        while True:
-            all_files = [file for file in self._temp_dir.iterdir() if file.suffix == ".json"]
-
-            song_choice_list = [{"name": file.stem, "value": file} for file in all_files]
-
-            if not song_choice_list:
-                song_choice_list.append(Choice(title="No songs available, please download a song first", disabled=True))
-
-            user_song_choice = questionary_select("Please choose the song:",
-                                                   choose_data=song_choice_list,
-                                                   enable_pages=True,
-                                                   batch_data=True,
-                                                   batch_size=10,
-                                                   extra_navigation_options=[Choice("Download New", value="__new__", shortcut_key="n")])
-            if user_song_choice == "__new__":
-                from backend_new.extractors.downloader import Downloader
-
-                with Downloader() as downloader:
-                    downloader.query_song_spotify()
-            else:
-                break
-
-        song_data = read_json_file(user_song_choice)
-        song_context = SongContext(song_data, user_song_choice)
-        # endregion
-
-        # region choice list for skip processes, update when adding new processes
-        skip_options = self._config_json.get("skip_processes")
-
-        choice_list = [Choice("Download song",
-                               value="download_song",
-                               checked=skip_options['download_song'],
-                               disabled="Song already downloaded" if song_context.json_song_data.get("pre_processing", {}).get("downloaded") else None),
-                        Choice("Get Genius metadata",
-                               value="genius_metadata",
-                               checked=skip_options['genius_metadata'],
-                               disabled="Data already gathered" if song_context.json_song_data.get("genius_data") else None),
-                        Choice("Separate audio into stems",
-                               value="vocal_separation",
-                               checked=skip_options['vocal_separation'],
-                               disabled="Stems already separated" if song_context.json_song_data.get("vocal_separation", {}).get("separated") else None),
-                        Choice("Split and tag lyrics",
-                               value="split_and_tag",
-                               description="(morphological analysis)",
-                               checked=skip_options['split_and_tag'],
-                               disabled="Lyrics already split and tagged" if song_context.json_song_data.get("split_and_tag", None) else None),
-                        Choice("Translate lyrics to english",
-                               value="translate_lyrics",
-                               description="(LLM inference)",
-                               checked=skip_options['translate_lyrics'],
-                               disabled="Lyrics already translated" if song_context.json_song_data.get("translated_lyrics",None) else None)
-                        ]
-
-        skip_processes = questionary_checkbox("Please choose what options to skip", choice_data=choice_list)
-        if skip_processes != [k for k in skip_options if skip_options[k]]:
-            write_skip_to_json = confirm("Do you want to set this to default?").ask()
-        else:
-            write_skip_to_json = False
-
-        for key in skip_options:
-            if key in skip_processes:
-                skip_options[key] = True
-            else:
-                skip_options[key] = False
-
-        if write_skip_to_json:
-            write_json_file(self._config_file, skip_options, ["skip_processes"])
-
-
-        # endregion
-
-        # region main loop for processes
-        with WorkflowManager(song_context) as manager:
-            # download
-            if not skip_options["download_song"]:
-                manager.download(song_context)
-            update_song_data()
-
-            # genius metadata
-            if not skip_options["genius_metadata"]:
-                manager.genius_pull(song_context)
-            update_song_data()
-
-            # vocal separation
-            if not skip_options["vocal_separation"]:
-                if not song_context.json_song_data["pre_processing"].get("downloaded", False):
-                    logger.info("Audio cannot be separated as it hasn't been downloaded")
-                    if confirm("Would you like to download first?").ask():
-                        manager.download(song_context)
-                    else:
-                        logger.warning("Audio is not present, skipping vocal separation")
-                if song_context.json_song_data["pre_processing"].get("downloaded", False):
-                    manager.vocal_sep(song_context)
-
-            # split and tag
-            if not skip_options["split_and_tag"]:
-                if song_context.json_song_data.get("genius_data", None) is None:
-                    logger.info("Genius metadata not present")
-                    if confirm("Would you like to pull it now?").ask():
-                        manager.genius_pull(song_context)
-                    else:
-                        logger.warning("Genius data not present, skipping split and tag")
-                if song_context.json_song_data.get("genius_data", None) is not None:
-                    manager.lyrics_tag(song_context)
-
-            # translate lyrics
-            if not skip_options["translate_lyrics"]:
-                if song_context.json_song_data.get("genius_data", None) is None:
-                    logger.info("Genius metadata not present")
-                    if confirm("Would you like to pull it now?").ask():
-                        manager.genius_pull(song_context)
-                    else:
-                        logger.warning("Genius data not present, skipping translation of lyrics")
-                if song_context.json_song_data.get("genius_data", None) is not None:
-                    manager.translate_lyrics(song_context)
-        # endregion
 
 
 def main() -> None:
