+++ src/backend_new/tui/widgets/static.py
@@ -101,6 +101,10 @@ class SpotifyCurrentlyPlayingWidget(Widget):
         self.set_interval(0.1, self.increment_timestamp)
         self.update_timer = self.set_interval(2, self.update_data)
 
+    @property
+    def song_available(self) -> bool:
+        return bool(self.response)
+
     @work(thread=True)
     def action_authenticate(self) -> None:
         if not self.app.use_spotify_token:
